# Conflicts:
#	base_layer/core/src/base_node/sync/block_sync/synchronizer.rs
#	base_layer/core/src/chain_storage/blockchain_database.rs
#	base_layer/core/src/chain_storage/lmdb_db/lmdb_db.rs
#	base_layer/p2p/src/initialization.rs
#	comms/dht/examples/memory_net/utilities.rs
#	comms/dht/examples/memorynet.rs
#	comms/dht/src/broadcast_strategy.rs
#	comms/dht/src/config.rs
#	comms/dht/src/connectivity/mod.rs
#	comms/dht/src/network_discovery/test.rs
#	comms/dht/src/outbound/broadcast.rs
#	comms/src/peer_manager/manager.rs
#	comms/src/peer_manager/peer_storage.rs
#	docs/src/SUMMARY.md
#	integration_tests/helpers/baseNodeProcess.js

+++ infrastructure/storage/src/key_val_store/lmdb_database.rs
@@ -146,10 +146,10 @@ mod test {
             struct Foo {
                 value: String,
             }
-            let key1 = 1 as u64;
-            let key2 = 2 as u64;
-            let key3 = 3 as u64;
-            let key4 = 4 as u64;
+            let key1 = 1;
+            let key2 = 2;
+            let key3 = 3;
+            let key4 = 4;
             let val1 = Foo {
                 value: "one".to_string(),
             };
